% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping.R
\name{ffinemap}
\alias{ffinemap}
\title{Functional fine mapping}
\usage{
ffinemap(flfdr, fpi0)
}
\arguments{
\item{flfdr}{A vector of functional local FDRs of a region of interest}

\item{fpi0}{An estimate of the function proportion of null tests}
}
\value{
A list of object type "sffdr" containing:
\item{BF}{The functional local Bayes' factors.}
\item{PP}{Posterior probability of a SNP being causal.}
}
\description{
Perform functional fine mapping with a set of functional local FDRs in a region of interest (assuming a single causal variant).
}
