% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpvalues.R
\name{fpvalues}
\alias{fpvalues}
\title{Functional p-values}
\usage{
fpvalues(lfdr, p = NULL)
}
\arguments{
\item{lfdr}{A vector of functional local FDRs of a region}

\item{p}{A vector of p-values. Default is NULL.}
}
\value{
A list of object type "sffdr" containing:
\item{fp}{Functional p-values.}
\item{fq}{Functional q-values.}
}
\description{
Calculate functional p-values from functional local FDRs. Internal use.
}
