% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{monoSmooth_conditional}
\alias{monoSmooth_conditional}
\title{Fast Conditional Monotonic Smoothing}
\usage{
monoSmooth_conditional(pvalue, density, group)
}
\arguments{
\item{pvalue}{Numeric vector of p-values.}

\item{density}{Numeric vector of estimated densities.}

\item{group}{Integer vector of surrogate bins.}
}
\value{
A numeric vector of smoothed densities.
}
\description{
Enforces strictly non-increasing density conditional on surrogate bins.
Note: Data must be sorted by (group, pvalue) before passing.
}
