% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{monoSmooth_pava}
\alias{monoSmooth_pava}
\title{Fast Conditional PAVA (Decreasing)}
\usage{
monoSmooth_pava(pvalue, density, group)
}
\arguments{
\item{pvalue}{A numeric vector of p-values. Must be sorted ascendingly within each group.}

\item{density}{A numeric vector of initial density estimates (e.g., from locfit).}

\item{group}{An integer vector denoting the surrogate bin/stratum for each observation.}
}
\value{
A numeric vector of smoothed, monotonically non-increasing densities.
}
\description{
Fast Conditional PAVA (Decreasing)
}
