% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_overlap.R
\name{run_empirical_overlap_correction}
\alias{run_empirical_overlap_correction}
\title{Run Overlap Correction using Data-Driven Nulls}
\usage{
run_empirical_overlap_correction(z1, z2, p1, p2)
}
\arguments{
\item{z1}{Numeric vector; primary trait z-scores.}

\item{z2}{Numeric vector; surrogate trait z-scores.}

\item{p1}{Numeric vector; primary trait p-values.}

\item{p2}{Numeric vector; surrogate trait p-values.}
}
\value{
A list containing the estimated rho_o, the conditional f0 density,
  and the indices of the empirical nulls used.
}
\description{
Run Overlap Correction using Data-Driven Nulls
}
