% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sffdr-package.R
\docType{package}
\name{sffdr-package}
\alias{sffdr-package}
\title{sffdr: Surrogate Functional False Discovery Rates for Genome-Wide Association Studies}
\description{
Pleiotropy-informed significance analysis of genome-wide association studies with surrogate functional false discovery rates (sfFDR). The sfFDR framework adapts the fFDR to leverage informative data from multiple sets of GWAS summary statistics to increase power in study while accommodating for linkage disequilibrium. sfFDR provides estimates of key FDR quantities in a significance analysis such as the functional local FDR and $q$-value, and uses these estimates to derive a functional $p$-value for type I error rate control and a functional local Bayes' factor for post-GWAS analyses (e.g., fine mapping and colocalization).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ajbass/sffdr}
}

}
\author{
\strong{Maintainer}: Andrew Bass \email{ab3105@cam.ac.uk}

Authors:
\itemize{
  \item Chris Wallace \email{chris.wallace@cam.ac.uk}
}

}
\keyword{internal}
